[Chore/#20] Model 파일로 분리
